Development

+++ app/api/auth_routes.py
@@ -87,11 +87,10 @@ def update_user(id):
     if form.validate_on_submit():
         user = User.query.filter(User.id == id).first()
 
-
-        user.username=form.data['username']
-        user.email=form.data['email']
-        user.password=form.data['password']
-        user.avatar=form.data['avatar']
+        user.username = form.data['username']
+        user.email = form.data['email']
+        user.password = form.data['password']
+        user.avatar = form.data['avatar']
 
         db.session.commit()
         return user.to_dict()
@@ -104,10 +103,13 @@ def delete_user(id):
     """
     Deletes a user account
     """
-    user = User.query.filter(User.id == id)
-    user.delete()
-    db.session.commit()
-    return dict(user)
+    user = User.query.get(id)
+    if user.id == 1:
+        return {'error': ['You cannot delete the demo user.']}, 403
+    else:
+        user.delete()
+        db.session.commit()
+        return dict(user)
 
 
 @auth_routes.route('/unauthorized')
